+++ source/components/FormField.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import SkinnableComponent from './SkinnableComponent';
+import { LabelProp } from "../utils/props";
 
 export default class FormField extends SkinnableComponent {
 
   static propTypes = Object.assign({}, SkinnableComponent.propTypes, {
     skin: PropTypes.element.isRequired,
-    label: PropTypes.string,
+    label: LabelProp,
     onChange: PropTypes.func,
     onFocus: PropTypes.func,
     onBlur: PropTypes.func,